:truck: [repositories] Rename domain.{urls => locations}
