+++ src/cutty/repositories/domain/locations.py
